Adds routes folder and server router file, adds twitter bot into own file
